+++ src/test/java/db/TransferTest.java
@@ -80,8 +80,8 @@ public class TransferTest extends BaseTest {
         BankWidget.dodepInAccount(user1.getUsername(), user1.getPassword(), userAccounts.get(0).getId(), balance);
 
         TransferRequestDTO transferRequestDTO = TransferRequestDTO.builder()
-                .senderAccountId(userAccounts.get(0).getId())
-                .receiverAccountId(userAccounts.get(receiverAccountId).getId())
+                .senderAccountId((long) userAccounts.get(0).getId())
+                .receiverAccountId((long) userAccounts.get(receiverAccountId).getId())
                 .amount(amount)
                 .build();
 
@@ -113,8 +113,8 @@ public class TransferTest extends BaseTest {
         BankWidget.dodepInAccount(user1.getUsername(), user1.getPassword(), userAccounts.get(0).getId(), balance);
 
         TransferRequestDTO transferRequestDTO = TransferRequestDTO.builder()
-                .senderAccountId(userAccounts.get(0).getId())
-                .receiverAccountId(userAccounts.get(receiverAccountId).getId())
+                .senderAccountId((long) userAccounts.get(0).getId())
+                .receiverAccountId((long) userAccounts.get(receiverAccountId).getId())
                 .amount(amount)
                 .build();
 
@@ -146,8 +146,8 @@ public class TransferTest extends BaseTest {
         BankWidget.dodepInAccount(user1.getUsername(), user1.getPassword(), userAccounts.get(0).getId(), balance);
 
         TransferRequestDTO transferRequestDTO = TransferRequestDTO.builder()
-                .senderAccountId(userAccounts.get(0).getId())
-                .receiverAccountId(0)
+                .senderAccountId((long) userAccounts.get(0).getId())
+                .receiverAccountId(0L)
                 .amount(balance)
                 .build();
